requirements.txt
flask
flask-socketio
flask-cors
python-dotenv
openai
email-validator
pytest
pytest-asyncio
requests
imaplib2